refactored code_added three classes_did homework
